Separate tokens, credentials and backdrop config
I had previously lumped this together under the concept of a 'service'.
It realised it was silly to lump all this information together as we
should have one source of backdrop config, the tokens will need to vary
independant of the creds for external users and there should only be one
set of creds for a 3rd party API.

The cronjobs now include a reference to query, creds, token and backdrop
config files.

The token stored in the query references one of the files in tokens/. I
put it in the query so that we can carry on auto generating the crontabs
file.

You may see some query files change more drastically than other or have
been removed. This is where I had hand converted queries and either got
the path wrong (in the case of those deleted) or the json has been re
ordered by convert.py